got insert to work
